Trim points on each oscilloscope_cache:process call to configured bounds
If points for a given metric compress well, it's likely that more than
(Interval * Count) points will be stored in-memory for that metric. This
is simply a waste of memory, since its exceeding the contract we make
with the user. This patch corrects this behavior, preventing stored points
from ever exceeding the configured bounds.