Fix string/binary comparison bug in auth
When oscilloscope loads users from PSQL, it loads the password hash as a
binary. Prior to this commit, oscilloscope auth would compare the binary
password hash and list password hash and therefore always 401 (if the
hash had been loaded from PSQL). This makes the comparison always happen
as a binary.