Immediately retry persist completion in case of redis timeout
After a successful persistent store write it's possible that the system
will fail while updating the in-memory cache, since that is currently
stored in Redis. This is an issue, because after the cache crashes it will
immediately attempt to persist the same data again. This second persist
will encounter one of two situations:

1. If the data chunks persisted are identical to the previous attempt, the
   new persist attempt will succeed, but will fail when attempting to
   insert persist records in to Postgres. This will loop until it hits case
   number 2.

2. If the data chunks persisted differ from the previous attempt, both the
   persist and the Postgres edits will succeed and the cache will
   successfully have saved data points in overlapping chunks. There's no
   invariant to verify that this doesn't happen, and the behavior of a
   cache in that state is unknown, at least to me.

This patch takes the simplest approach and causes any failure in the persist
finalization attempts to be retried as soon as possible.