Use a converging "guess" to estimate necessary chunk size
The previous implementation failed to converge in some secnarios, resulting in
an infinite loop.

This implementation attempts to converge faster by continually improving a
"guess" at the number of items necessary to form a complete chunk.

This could be further improved by manually keeping track of the number of items
in the Values/Excess lists - we call length/0 on those quite a lot.

Furthermore, we should consider keeping track of infinite-loop-ness in the
persistence server. That behavior might be allevated by reducing min_chunk_size.

Finally, it seems that this would be a great candidate for some PropEr-style
tests.