Begin re-architecture on top of riak_core
This commit is a first step in replacing the existing single-node Oscilloscope
architecture with one built on top of riak_core. The ultimate goal of this
process is horizontal scalability, however this commit alone is insufficient.

There are several meaningful architectural changes here that I'd like to point
out:

1. Metrics are not mapped on a 1:1 basis with Erlang processes. Instead, metrics
   are hashed across a set of riak_core vnodes, where each vnode handles data
   for total_metrics/ring_size metrics.

2. Cached data management is now isolated to a library module that encompasses
   all retention schemas for a given metric. This module is TODO.

3. The data caches are completely decoupled from the persistence application;
   the logic to read data from each source now exists in the oscilloscope
   action, where a gen_fsm will handle reading data from the various
   sources (cached/persisted) as necessary.

Much of the above is TODO; this is a WIP commit that lays down the basic
skeleton of the application.