Make oscilloscope read from kafka
This commit adds the first version of oscilloscope which reads data from
kafka. Each instance of oscilloscope is responsible for reading from a
set of partitions. A process (see: `osc_kafka_partition.erl`) is spawned
for each partition which is owned by the node. The partition proc reads
data from the partition and writes metrics to the caches.