Replace erlang-bcrypt with pbkdf2
Prior to this patch, password-based authentication was very slow. This was due
to two reasons:

1. The default work factor in erlang-bcrypt is 12, which is quite high. This is,
of course, a manageable problem - we could just reduce the work factor and move
forward.

2. erlang-bcrypt is single-threaded (in Erlang!) and the version we were using
was not thread-safe. Every password hash was processed by a single gen_sever,
which resulted in painful effects on actual graph load times when, e.g.,
Graphite would send 20 simultaneous queries at the system.

In spite of the apparent "simplicity" of the erlang-bcrypt project, it's also a
bit of a mess. There's an async thread pool in the NIF and callers sit in a
receive block waiting for bcrypt to do its work, even though the app is
single-threaded.

This patch replaces all of that complexity with an Erlang implementation of
pbkdf2 based on rnewson's work in CouchDB. The major deviation from the CouchDB
implementation is the use of sha512 instead of sha1. I took the liberty of
cleaning up the code a bit as well.