Handle disjoint and overlapping read windows
Prior to this patch a window that had non-continuous data persisted - e.g.,
there is data in the persistent store for t = {0, 10, 20, 90, 100, 110} with
interval = 10 - would break the read logic, because osc_persistence:read_int/3
implicitly assumed a continuous stream of points. It just joined them all
together.

This patch treats every set of read points as an independent "read" in the
merging process, thereby both simplifying the code and making it gracefully
handle disjoint windows.

Furthermore, this patch will gracefully handle overlapping windows and log them
when they occur. The system *should* guarantee that these aren't possible, but a
bug in the 0.0.2 era allowed them - we're not guaranteed to never see them.